Rename domain pkg to dto
