+++ src/main/java/org/example/dto/parameters/BeddingMaterial.java
@@ -1,4 +1,4 @@
-package org.example.domain.parameters;
+package org.example.dto.parameters;
 
 public enum BeddingMaterial {
     COTTON,
